main menu
basic ui elements and view switching

+++ src/main/game/entities/Player.java
@@ -39,7 +39,7 @@ public class Player extends Entity implements WeaponController {
 
         terrainCollisionRadius = radius;
 
-        sprite = new Sprite(SpriteData.PLAYER, x, y, radius);
+        sprite = new Sprite(SpriteData.PLAYER, (float)x, (float)y, (float)radius);
 
         hitbox = new BodyHitbox(this, team, radius) {
             @Override
@@ -99,9 +99,9 @@ public class Player extends Entity implements WeaponController {
     public void updatePost(double delta) {
         x = nextX;
         y = nextY;
-        sprite.setRotation(sprite.getRotation() + 0.01);
+        sprite.setRotation(sprite.getRotation() + 0.01f);
         sprite.setImageIndex((int)(WindowManager.getTime() * 1.0) % 4);
-        sprite.setPosition(x, y);
+        sprite.setPosition((float)x, (float)y);
 
         hitbox.setPosition(x, y);
